Initial work on problem before refactoring.
This is the version with “factories”. Factory Girl isn’t really required
as Object.new works pretty much like a factory in and of itself but my
thoughts were that people are coming to a talk on factories and not
using Factory Girl may be more confusing than using it. I may change my
mind though as it’s simpler without, and it may hinder the understanding
of those who aren’t familiar with Factory Girl.